Add `git tug` alias
When adding a tag to a commit/repo, the usual pattern is to have the
tag version on both the tag's name and tag's message like:

```bash
git tag -a v0.8.11 -m "v0.8.11"
```

This poses a risk as re-using the same command (by using reverse
search for instance), one might end up changing only the message to
the latest tag version, but not the tag version, essentially messing
things up.

Let's automate this process by creating an alias that requires only
one version to be written out. The alias will take care of duplicating
the version to both tag's name and message.