Fix NERDTree's `hit-enter`, update rust-analyzer
We fix NERDTree's sudden issue related to CWD and `hit-enter`.
NERDTree updates the CWD to the tab's CWD everytime user moves to
another tab. However, when the CWD is too long, the `hit-enter`
command prompt shows up which is annoying. We fix that by leveraging
vim's `messagesopt`: whenever user goes to the command line we add
the `hit-enter` there, but whenever user changes tab we remove it. If
we had used the `CmdLineLeave` event, it would be impossible for the
user to make use of any `messages` buffer etc.

We also update the `rust-analyzer` for vim9 by removing the default
hints config.